# backend/api/gen/siem/v1/admin_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: siem/v1/admin.proto

package siemv1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationAdminCreateUser = "/siem.v1.Admin/CreateUser"
const OperationAdminDeleteUser = "/siem.v1.Admin/DeleteUser"
const OperationAdminGetCorrelationSettings = "/siem.v1.Admin/GetCorrelationSettings"
const OperationAdminGetTenantSettings = "/siem.v1.Admin/GetTenantSettings"
const OperationAdminListAuditEntries = "/siem.v1.Admin/ListAuditEntries"
const OperationAdminListUsers = "/siem.v1.Admin/ListUsers"
const OperationAdminPurge = "/siem.v1.Admin/Purge"
const OperationAdminUpdateCorrelationSettings = "/siem.v1.Admin/UpdateCorrelationSettings"
const OperationAdminUpdateTenantSettings = "/siem.v1.Admin/UpdateTenantSettings"
const OperationAdminUpdateUser = "/siem.v1.Admin/UpdateUser"

type AdminHTTPServer interface {
	CreateUser(context.Context, *CreateUserRequest) (*UserProfile, error)
	DeleteUser(context.Context, *DeleteUserRequest) (*DeleteUserResponse, error)
	GetCorrelationSettings(context.Context, *GetCorrelationSettingsRequest) (*CorrelationSettings, error)
	GetTenantSettings(context.Context, *GetTenantSettingsRequest) (*TenantSettings, error)
	// ListAuditEntries Read-only. There is deliberately no operation that writes, edits, or deletes an
	// audit entry.
	ListAuditEntries(context.Context, *ListAuditEntriesRequest) (*ListAuditEntriesResponse, error)
	ListUsers(context.Context, *ListUsersRequest) (*ListUsersResponse, error)
	// Purge Explicit, audited data deletion outside the normal retention path.
	Purge(context.Context, *PurgeRequest) (*PurgeResponse, error)
	UpdateCorrelationSettings(context.Context, *UpdateCorrelationSettingsRequest) (*CorrelationSettings, error)
	// UpdateTenantSettings Retention and redaction changes are audited with before and after values.
	UpdateTenantSettings(context.Context, *UpdateTenantSettingsRequest) (*TenantSettings, error)
	UpdateUser(context.Context, *UpdateUserRequest) (*UserProfile, error)
}

func RegisterAdminHTTPServer(s *http.Server, srv AdminHTTPServer) {
	r := s.Route("/")
	r.GET("/api/v1/admin/users", _Admin_ListUsers0_HTTP_Handler(srv))
	r.POST("/api/v1/admin/users", _Admin_CreateUser0_HTTP_Handler(srv))
	r.PATCH("/api/v1/admin/users/{user_id}", _Admin_UpdateUser0_HTTP_Handler(srv))
	r.DELETE("/api/v1/admin/users/{user_id}", _Admin_DeleteUser0_HTTP_Handler(srv))
	r.GET("/api/v1/admin/tenant", _Admin_GetTenantSettings0_HTTP_Handler(srv))
	r.PATCH("/api/v1/admin/tenant", _Admin_UpdateTenantSettings0_HTTP_Handler(srv))
	r.GET("/api/v1/admin/correlation-settings", _Admin_GetCorrelationSettings0_HTTP_Handler(srv))
	r.PATCH("/api/v1/admin/correlation-settings", _Admin_UpdateCorrelationSettings0_HTTP_Handler(srv))
	r.POST("/api/v1/admin/purge", _Admin_Purge0_HTTP_Handler(srv))
	r.GET("/api/v1/audit", _Admin_ListAuditEntries0_HTTP_Handler(srv))
}

func _Admin_ListUsers0_HTTP_Handler(srv AdminHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListUsersRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAdminListUsers)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListUsers(ctx, req.(*ListUsersRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListUsersResponse)
		return ctx.Result(200, reply)
	}
}

func _Admin_CreateUser0_HTTP_Handler(srv AdminHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateUserRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAdminCreateUser)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateUser(ctx, req.(*CreateUserRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UserProfile)
		return ctx.Result(200, reply)
	}
}

func _Admin_UpdateUser0_HTTP_Handler(srv AdminHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateUserRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAdminUpdateUser)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateUser(ctx, req.(*UpdateUserRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UserProfile)
		return ctx.Result(200, reply)
	}
}

func _Admin_DeleteUser0_HTTP_Handler(srv AdminHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteUserRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAdminDeleteUser)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteUser(ctx, req.(*DeleteUserRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteUserResponse)
		return ctx.Result(200, reply)
	}
}

func _Admin_GetTenantSettings0_HTTP_Handler(srv AdminHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetTenantSettingsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAdminGetTenantSettings)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetTenantSettings(ctx, req.(*GetTenantSettingsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*TenantSettings)
		return ctx.Result(200, reply)
	}
}

func _Admin_UpdateTenantSettings0_HTTP_Handler(srv AdminHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateTenantSettingsRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAdminUpdateTenantSettings)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateTenantSettings(ctx, req.(*UpdateTenantSettingsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*TenantSettings)
		return ctx.Result(200, reply)
	}
}

func _Admin_GetCorrelationSettings0_HTTP_Handler(srv AdminHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetCorrelationSettingsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAdminGetCorrelationSettings)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetCorrelationSettings(ctx, req.(*GetCorrelationSettingsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CorrelationSettings)
		return ctx.Result(200, reply)
	}
}

func _Admin_UpdateCorrelationSettings0_HTTP_Handler(srv AdminHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateCorrelationSettingsRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAdminUpdateCorrelationSettings)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateCorrelationSettings(ctx, req.(*UpdateCorrelationSettingsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CorrelationSettings)
		return ctx.Result(200, reply)
	}
}

func _Admin_Purge0_HTTP_Handler(srv AdminHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in PurgeRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAdminPurge)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Purge(ctx, req.(*PurgeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*PurgeResponse)
		return ctx.Result(200, reply)
	}
}

func _Admin_ListAuditEntries0_HTTP_Handler(srv AdminHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListAuditEntriesRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAdminListAuditEntries)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListAuditEntries(ctx, req.(*ListAuditEntriesRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListAuditEntriesResponse)
		return ctx.Result(200, reply)
	}
}

type AdminHTTPClient interface {
	CreateUser(ctx context.Context, req *CreateUserRequest, opts ...http.CallOption) (rsp *UserProfile, err error)
	DeleteUser(ctx context.Context, req *DeleteUserRequest, opts ...http.CallOption) (rsp *DeleteUserResponse, err error)
	GetCorrelationSettings(ctx context.Context, req *GetCorrelationSettingsRequest, opts ...http.CallOption) (rsp *CorrelationSettings, err error)
	GetTenantSettings(ctx context.Context, req *GetTenantSettingsRequest, opts ...http.CallOption) (rsp *TenantSettings, err error)
	// ListAuditEntries Read-only. There is deliberately no operation that writes, edits, or deletes an
	// audit entry.
	ListAuditEntries(ctx context.Context, req *ListAuditEntriesRequest, opts ...http.CallOption) (rsp *ListAuditEntriesResponse, err error)
	ListUsers(ctx context.Context, req *ListUsersRequest, opts ...http.CallOption) (rsp *ListUsersResponse, err error)
	// Purge Explicit, audited data deletion outside the normal retention path.
	Purge(ctx context.Context, req *PurgeRequest, opts ...http.CallOption) (rsp *PurgeResponse, err error)
	UpdateCorrelationSettings(ctx context.Context, req *UpdateCorrelationSettingsRequest, opts ...http.CallOption) (rsp *CorrelationSettings, err error)
	// UpdateTenantSettings Retention and redaction changes are audited with before and after values.
	UpdateTenantSettings(ctx context.Context, req *UpdateTenantSettingsRequest, opts ...http.CallOption) (rsp *TenantSettings, err error)
	UpdateUser(ctx context.Context, req *UpdateUserRequest, opts ...http.CallOption) (rsp *UserProfile, err error)
}

type AdminHTTPClientImpl struct {
	cc *http.Client
}

func NewAdminHTTPClient(client *http.Client) AdminHTTPClient {
	return &AdminHTTPClientImpl{client}
}

func (c *AdminHTTPClientImpl) CreateUser(ctx context.Context, in *CreateUserRequest, opts ...http.CallOption) (*UserProfile, error) {
	var out UserProfile
	pattern := "/api/v1/admin/users"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationAdminCreateUser))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *AdminHTTPClientImpl) DeleteUser(ctx context.Context, in *DeleteUserRequest, opts ...http.CallOption) (*DeleteUserResponse, error) {
	var out DeleteUserResponse
	pattern := "/api/v1/admin/users/{user_id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationAdminDeleteUser))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *AdminHTTPClientImpl) GetCorrelationSettings(ctx context.Context, in *GetCorrelationSettingsRequest, opts ...http.CallOption) (*CorrelationSettings, error) {
	var out CorrelationSettings
	pattern := "/api/v1/admin/correlation-settings"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationAdminGetCorrelationSettings))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *AdminHTTPClientImpl) GetTenantSettings(ctx context.Context, in *GetTenantSettingsRequest, opts ...http.CallOption) (*TenantSettings, error) {
	var out TenantSettings
	pattern := "/api/v1/admin/tenant"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationAdminGetTenantSettings))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListAuditEntries Read-only. There is deliberately no operation that writes, edits, or deletes an
// audit entry.
func (c *AdminHTTPClientImpl) ListAuditEntries(ctx context.Context, in *ListAuditEntriesRequest, opts ...http.CallOption) (*ListAuditEntriesResponse, error) {
	var out ListAuditEntriesResponse
	pattern := "/api/v1/audit"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationAdminListAuditEntries))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *AdminHTTPClientImpl) ListUsers(ctx context.Context, in *ListUsersRequest, opts ...http.CallOption) (*ListUsersResponse, error) {
	var out ListUsersResponse
	pattern := "/api/v1/admin/users"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationAdminListUsers))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Purge Explicit, audited data deletion outside the normal retention path.
func (c *AdminHTTPClientImpl) Purge(ctx context.Context, in *PurgeRequest, opts ...http.CallOption) (*PurgeResponse, error) {
	var out PurgeResponse
	pattern := "/api/v1/admin/purge"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationAdminPurge))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *AdminHTTPClientImpl) UpdateCorrelationSettings(ctx context.Context, in *UpdateCorrelationSettingsRequest, opts ...http.CallOption) (*CorrelationSettings, error) {
	var out CorrelationSettings
	pattern := "/api/v1/admin/correlation-settings"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationAdminUpdateCorrelationSettings))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PATCH", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateTenantSettings Retention and redaction changes are audited with before and after values.
func (c *AdminHTTPClientImpl) UpdateTenantSettings(ctx context.Context, in *UpdateTenantSettingsRequest, opts ...http.CallOption) (*TenantSettings, error) {
	var out TenantSettings
	pattern := "/api/v1/admin/tenant"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationAdminUpdateTenantSettings))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PATCH", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *AdminHTTPClientImpl) UpdateUser(ctx context.Context, in *UpdateUserRequest, opts ...http.CallOption) (*UserProfile, error) {
	var out UserProfile
	pattern := "/api/v1/admin/users/{user_id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationAdminUpdateUser))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PATCH", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}