# backend/api/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /api/v1/admin/correlation-settings:
        get:
            tags:
                - Admin
            operationId: Admin_GetCorrelationSettings
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CorrelationSettings'
        patch:
            tags:
                - Admin
            operationId: Admin_UpdateCorrelationSettings
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateCorrelationSettingsRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CorrelationSettings'
    /api/v1/admin/purge:
        post:
            tags:
                - Admin
            description: Explicit, audited data deletion outside the normal retention path.
            operationId: Admin_Purge
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PurgeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PurgeResponse'
    /api/v1/admin/tenant:
        get:
            tags:
                - Admin
            operationId: Admin_GetTenantSettings
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TenantSettings'
        patch:
            tags:
                - Admin
            description: Retention and redaction changes are audited with before and after values.
            operationId: Admin_UpdateTenantSettings
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateTenantSettingsRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TenantSettings'
    /api/v1/admin/users:
        get:
            tags:
                - Admin
            operationId: Admin_ListUsers
            parameters:
                - name: page.cursor
                  in: query
                  description: Opaque cursor from a previous PageResponse. Empty requests the first page.
                  schema:
                    type: string
                - name: page.limit
                  in: query
                  description: Server caps this at QUERY_MAX_RESULT_ROWS regardless of what is requested.
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListUsersResponse'
        post:
            tags:
                - Admin
            operationId: Admin_CreateUser
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateUserRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserProfile'
    /api/v1/admin/users/{userId}:
        delete:
            tags:
                - Admin
            operationId: Admin_DeleteUser
            parameters:
                - name: userId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteUserResponse'
        patch:
            tags:
                - Admin
            operationId: Admin_UpdateUser
            parameters:
                - name: userId
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateUserRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserProfile'
...
    /api/v1/audit:
        get:
            tags:
                - Admin
            description: |-
                Read-only. There is deliberately no operation that writes, edits, or deletes an
                 audit entry.
            operationId: Admin_ListAuditEntries
            parameters:
                - name: range.from
                  in: query
                  schema:
                    type: string
                    format: date-time
                - name: range.to
                  in: query
                  schema:
                    type: string
                    format: date-time
                - name: page.cursor
                  in: query
                  description: Opaque cursor from a previous PageResponse. Empty requests the first page.
                  schema:
                    type: string
                - name: page.limit
                  in: query
                  description: Server caps this at QUERY_MAX_RESULT_ROWS regardless of what is requested.
                  schema:
                    type: integer
                    format: int32
                - name: action
                  in: query
                  schema:
                    type: string
                - name: actorEmail
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListAuditEntriesResponse'
...
components:
    schemas:
...
        AuditEntry:
            type: object
            properties:
                entryId:
                    type: string
                occurredAt:
                    type: string
                    format: date-time
                actorEmail:
                    type: string
                sourceIp:
                    type: string
                action:
                    type: string
                targetType:
                    type: string
                targetId:
                    type: string
                beforeValue:
                    type: string
                afterValue:
                    type: string
                result:
                    type: string
                detail:
                    type: string
                entryHash:
                    type: string
                    description: |-
                        The hash chain. Exposed so a reader can verify the LINKAGE themselves rather than
                         take the platform's word for it — an audit trail whose integrity can only be
                         confirmed by the system that wrote it is not evidence.

                         Checking these proves no entry was removed from the middle of a range: a deletion
                         breaks the chain where it happened. It does NOT prove the whole range was not
                         rewritten from the start; that needs the server-held anchor.
                previousHash:
                    type: string
...
        CorrelationSettings:
            type: object
            properties:
                correlationWindowMs:
                    type: integer
                    description: How far apart two events may be and still describe the same request.
                    format: uint32
                latenessBoundMs:
                    type: integer
                    description: How late an event may arrive and still amend an existing correlated record.
                    format: uint32
                signalRanking:
                    type: array
                    items:
                        enum:
                            - JOIN_SIGNAL_UNSPECIFIED
                            - JOIN_SIGNAL_VENDOR_REQUEST_ID
                            - JOIN_SIGNAL_IP_HOST_PATH_METHOD
                            - JOIN_SIGNAL_TIME_WINDOW
                        type: string
                        format: enum
                    description: Ranked join signals, most trusted first.
...
        CreateUserRequest:
            type: object
            properties:
                email:
                    type: string
                password:
                    type: string
                role:
                    type: string
...
        DeleteUserResponse:
            type: object
            properties: {}
...
        ListAuditEntriesResponse:
            type: object
            properties:
                entries:
                    type: array
                    items:
                        $ref: '#/components/schemas/AuditEntry'
                page:
                    $ref: '#/components/schemas/PageResponse'
                chainIntact:
                    type: boolean
                    description: |-
                        False when the hash chain over the returned range does not verify, meaning an
                         entry was altered or removed.
...
        ListUsersResponse:
            type: object
            properties:
                users:
                    type: array
                    items:
                        $ref: '#/components/schemas/UserProfile'
                page:
                    $ref: '#/components/schemas/PageResponse'
...
        PageResponse:
            type: object
            properties:
                nextCursor:
                    type: string
                    description: Empty when there are no further pages.
                total:
                    type: string
                    description: |-
                        Approximate match count. Exact counts over hundreds of millions of rows cannot
                         meet the latency budget, so the API states when a count is an estimate rather
                         than implying precision it does not have.
                totalIsEstimate:
                    type: boolean
...
        PurgeRequest:
            type: object
            properties:
                range:
                    $ref: '#/components/schemas/TimeRange'
                confirmation:
                    type: string
                    description: |-
                        Must equal the tenant name. A destructive operation should not be reachable by
                         an accidental click.
                reason:
                    type: string
        PurgeResponse:
            type: object
            properties:
                rowsDeleted:
                    type: string
...
        TenantSettings:
            type: object
            properties:
                tenantId:
                    type: string
                name:
                    type: string
                status:
                    type: string
                rawRetentionDays:
                    type: integer
                    format: uint32
                correlatedRetentionDays:
                    type: integer
                    format: uint32
                alertRetentionDays:
                    type: integer
                    format: uint32
                redactedFields:
                    type: array
                    items:
                        type: string
                    description: Fields masked at ingest, so they are never stored in readable form.
                scoreConflictThreshold:
                    type: number
                    format: float
...
        TimeRange:
            type: object
            properties:
                from:
                    type: string
                    format: date-time
                to:
                    type: string
                    format: date-time
            description: |-
                TimeRange bounds a query. Both ends are REQUIRED on every event-data read:
                 unbounded scans are rejected, not queued (Constitution Principle V).
...
        UpdateCorrelationSettingsRequest:
            type: object
            properties:
                correlationWindowMs:
                    type: integer
                    format: uint32
                latenessBoundMs:
                    type: integer
                    format: uint32
                signalRanking:
                    type: array
                    items:
                        enum:
                            - JOIN_SIGNAL_UNSPECIFIED
                            - JOIN_SIGNAL_VENDOR_REQUEST_ID
                            - JOIN_SIGNAL_IP_HOST_PATH_METHOD
                            - JOIN_SIGNAL_TIME_WINDOW
                        type: string
                        format: enum
...
        UpdateTenantSettingsRequest:
            type: object
            properties:
                rawRetentionDays:
                    type: integer
                    format: uint32
                correlatedRetentionDays:
                    type: integer
                    format: uint32
                alertRetentionDays:
                    type: integer
                    format: uint32
                redactedFields:
                    type: array
                    items:
                        type: string
                scoreConflictThreshold:
                    type: number
                    format: float
        UpdateUserRequest:
            type: object
            properties:
                userId:
                    type: string
                role:
                    type: string
                status:
                    type: string
                resetMfa:
                    type: boolean
                    description: Setting this resets MFA enrolment, forcing re-enrolment on next login.
        UserProfile:
            type: object
            properties:
                userId:
                    type: string
                email:
                    type: string
                role:
                    type: string
                    description: admin | analyst | auditor | ingest_only
                tenantId:
                    type: string
                tenantName:
                    type: string
                mfaEnabled:
                    type: boolean
                lastLoginAt:
                    type: string
                    format: date-time
...